Add files via upload
CAD File for use with waterjet. All waterjet files are in the folder labelled .dxf First file to cut is FlowJet Test Cut. If this file creates a square that is 45 mm x 50 mm and has a hole that will accept our disk magnets as a press fit, then we are good to waterjet the scalene_chassis_basentop.dxf file out of HDPE, which will be our top and base plate. Rest of the files in this folder were used to waterjet the transmission plates from aluminum.